Update the style in the html file
Expand tabs to two spaces, limit the length of lines to 79 characters
and introduce new lines.

One unfortunate convention that had to be addopted is not having the
closing </a> tags on a separate line. This is done in order to avoid the
whitespace added if </a> is followed by a punctuation mark.